requirements.txt
web3>=6.0.0
telethon>=1.34.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
eth-abi>=5.0.0
